[INTERNAL] Split auth routes into several files
