+++ collectives/routes/auth/__init__.py
@@ -1,0 +1,7 @@
+from collectives.routes.auth.globals import login_manager, blueprint
+from collectives.routes.auth.utils import get_bad_phone_message, sync_user
+
+import collectives.routes.auth.utils
+import collectives.routes.auth.globals
+import collectives.routes.auth.login
+import collectives.routes.auth.signup
