+++ collectives/routes/auth/signup.py
@@ -1,89 +1,25 @@
-""" Module for user authentification routes. """
-
-import datetime
-from urllib.parse import urlparse
 
 from flask import flash, render_template, redirect, url_for, request
-from flask import current_app, Blueprint
-from markupsafe import Markup, escape
-from flask_login import current_user, login_user, logout_user, login_required
-from flask_login import LoginManager, AnonymousUserMixin
-from flask_wtf.csrf import generate_csrf
+from flask import current_app
+from flask_login import current_user
+from markupsafe import Markup
 from sqlalchemy import or_
 
-from collectives import email_templates
-from collectives.forms.auth import LoginForm, AccountCreationForm
-from collectives.forms.auth import PasswordResetForm, AccountActivationForm
-from collectives.models import User, db, Configuration, UserType
-from collectives.models.auth import ConfirmationTokenType, ConfirmationToken
-from collectives.models.auth import TokenEmailStatus
-from collectives.utils import extranet
+
+
+from collectives.routes.auth.globals import blueprint
+from collectives.models import db, ConfirmationToken, ConfirmationTokenType
+from collectives.models import User, Configuration
 from collectives.utils.time import current_time
+from collectives.forms.auth import PasswordResetForm, AccountActivationForm
+from collectives.forms.auth import AccountCreationForm
+from collectives.utils import extranet
+from collectives import email_templates
+from collectives.models.auth import TokenEmailStatus
 
 
-class UnauthenticatedUserMixin(AnonymousUserMixin):
-    """Mixin that defines properties for unauthenticated users"""
-
-    @property
-    def id(self) -> int:
-        """Id for unauthenticated users (always `-1`)"""
-        return -1
 
 
-login_manager = LoginManager()
-login_manager.anonymous_user = UnauthenticatedUserMixin
-login_manager.login_view = "auth.login"
-login_manager.login_message = "Merci de vous connecter pour accéder à cette page"
-
-blueprint = Blueprint("auth", __name__, url_prefix="/auth")
-""" Authentification blueprint
-
-This blueprint contains all routes for authentification actions.
-"""
-
-
-# Flask-login user loader
-@login_manager.user_loader
-def load_user(user_id):
-    """Flask-login user loader.
-
-    See also: `flask_login.LoginManager.user_loader
-    <https://flask-login.readthedocs.io/en/latest/#flask_login.LoginManager.user_loader>`_
-
-    :param string user_id: primary of the user in sql
-    :return: current user or None
-    :rtype: :py:class:`collectives.models.user.User`
-    """
-    user = db.session.get(User, int(user_id))
-    if user is None or not user.is_active:
-        # License has expired, log-out user
-        return None
-    return user
-
-
-def sync_user(user, force):
-    """Synchronize user info from extranet.
-
-    Synchronization is done if license has been renewed or if 'force' is True. Test users
-    cannot be synchronized.
-
-    :param user: User to synchronize
-    :type user: :py:class:`collectives.models.user.User`
-    :param force: if True, do synchronisation even if licence has been recently renewed.
-    :type force: boolean
-    """
-    if user.enabled and user.type == UserType.Extranet:
-        # Check whether the license has been renewed
-        license_info = extranet.api.check_license(user.license)
-        if not license_info.exists:
-            return
-
-        if force or license_info.expiry_date() > user.license_expiry_date:
-            # License has been renewd, sync user data from API
-            user_info = extranet.api.fetch_user_info(user.license)
-            extranet.sync_user(user, user_info, license_info)
-            db.session.add(user)
-            db.session.commit()
 
 
 def create_confirmation_token(license_number, user):
@@ -96,124 +32,6 @@ def create_confirmation_token(license_number, user):
     db.session.add(token)
     db.session.commit()
     return token
-
-
-@blueprint.route("/login", methods=["GET", "POST"])
-def login():
-    """Route for user login page.
-
-    If authentification has failed, there is a timeout : :py:data:`config.AUTH_FAILURE_WAIT`
-    If confidentiality agreement or legal text has not been signed, a reminder is
-    shown to user if login is succesful.
-    """
-    form = LoginForm()
-
-    # If no login is provided, display regular login interface
-    if not form.validate_on_submit():
-        return render_template(
-            "auth/login.html",
-            form=form,
-            contact_reason="vous connecter",
-        )
-
-    # Check if user exists
-    user = User.query.filter_by(mail=form.mail.data).first()
-
-    if user is None:
-        flash("Nom d'utilisateur ou mot de passe invalide.", "error")
-        return redirect(url_for("auth.login"))
-
-    maxdelta = datetime.timedelta(seconds=Configuration.AUTH_FAILURE_WAIT)
-    if current_time() - user.last_failed_login < maxdelta:
-        flash("Merci d'attendre quelques secondes avant de retenter un login", "error")
-        return redirect(url_for("auth.login"))
-
-    if not user.password == form.password.data:
-        user.last_failed_login = current_time()
-        db.session.add(user)
-        db.session.commit()
-        flash("Nom d'utilisateur ou mot de passe invalide.", "error")
-        return redirect(url_for("auth.login"))
-
-    try:
-        sync_user(user, False)
-    except extranet.ExtranetError:
-        flash(
-            """Synchronization avec l'extranet FFCAM impossible,
-              vos informations utilisateur pourront ne pas être à jour""",
-            "warning",
-        )
-
-    if not user.is_active:
-        flash(
-            Markup(
-                f"""Compte ou numéro de licence inactif, merci de renouveler votre adhésion.
-            Si vous avez changé de numéro de licence, utilisez le
-            <a href='{url_for("auth.recover")}'>formulaire de récupération de compte</a>."""
-            ),
-            "error",
-        )
-        return redirect(url_for("auth.login"))
-
-    login_user(user, remember=form.remember_me.data)
-
-    if not current_user.has_valid_phone_number():
-        flash(Markup(get_bad_phone_message(user)), "warning")
-    if not current_user.has_valid_phone_number(emergency=True):
-        flash(Markup(get_bad_phone_message(user, emergency=True)), "warning")
-
-    # We ask users with roles to sign the confidentiality agreement.
-    # Signature is compulsory to view user profiles.
-    if not user.has_signed_ca() and user.has_any_role():
-        url = url_for("profile.confidentiality_agreement")
-        flash(
-            Markup(
-                f"""Avec vos fonctions, vous pouvez accèder à
-                des informations personnelles d'adhérent.
-                Merci donc de signer la
-                <a href=\"{url}\">charte RGPD [ICI].</a>"""
-            ),
-            "warning",
-        )
-
-    if not user.has_signed_legal_text():
-        flash(Markup("""Merci d'accepter les mentions légales du site."""), "warning")
-        return redirect(url_for("root.legal"))
-
-    # Redirection to the page required by user before login
-    next_page = request.args.get("next")
-    if not next_page or urlparse(next_page).netloc != "":
-        next_page = "/"
-    return redirect(next_page)
-
-
-@blueprint.route("/logout")
-@login_required
-def logout():
-    """Route for logout user page."""
-    logout_user()
-    return redirect(url_for("auth.login"))
-
-
-def render_confirmation_form(form, is_recover):
-    """Render template with right options for email confirmation webpage.
-
-    :param form: Confirmation form which will be used to generate web page.
-    :type form: :py:class:`collectives.form.auth.PasswordResetForm` or
-                :py:class:`collectives.form.auth.AccountActivationForm`
-    :param is_recover: True if we are resetting the password.
-    :type is_recover: boolean
-    :return: Email confirmation webpage
-    """
-    action = "Récupération " if is_recover else "Création"
-    reason = "récupérer" if is_recover else "créer"
-    form.submit.label.text = f"{reason.capitalize()} le compte"
-    return render_template(
-        "auth/token_confirmation.html",
-        form=form,
-        title=f"{action} de compte",
-    )
-
 
 @blueprint.route("/process_confirmation/<token_uuid>", methods=["GET", "POST"])
 def process_confirmation(token_uuid):
@@ -460,40 +278,21 @@ def check_token(license_number):
 
     return redirect(url_for(".login"))
 
+def render_confirmation_form(form, is_recover):
+    """Render template with right options for email confirmation webpage.
 
-def get_bad_phone_message(user, emergency=False):
-    """Craft a message for user with a bad phone number.
-
-    :param user: User with a bad phone number
-    :param emergency: If true, generates a message relative to the emergency number
-    :returns: The html safe message to display to the user to help him change phone number.
+    :param form: Confirmation form which will be used to generate web page.
+    :type form: :py:class:`collectives.form.auth.PasswordResetForm` or
+                :py:class:`collectives.form.auth.AccountActivationForm`
+    :param is_recover: True if we are resetting the password.
+    :type is_recover: boolean
+    :return: Email confirmation webpage
     """
-    phone_number = user.emergency_contact_phone if emergency else user.phone
-
-    if phone_number is None:
-        phone = "n'est pas"
-    else:
-        phone = f"({escape(phone_number)}) est mal"
-
-    if not emergency:
-        description = """un encadrant ne pourrait donc pas vous contacter pour vous communiquer
-        des informations importantes comme un changement de lieu de rendez-vous, un changement 
-        du lieu de la sortie ou l'annulation de la sortie"""
-    else:
-        description = """le club ne pourrait donc pas contacter vos proches en cas de
-        besoin."""
-
-    return f"""Votre numéro de téléphone {"d'urgence " if emergency else ''} {phone} renseigné:
-        {description}<br/>
-        Veuillez saisir un numéro de téléphone valide dans <a 
-        href="https://extranet-clubalpin.com/monespace/">
-        votre espace personnel FFCAM </a>, menu "Mes informations", puis resynchronisez vos 
-        informations en cliquant 
-        <form action="{url_for('profile.force_user_sync')}" method="post" style="display: inline">
-            <a onclick="this.closest('form').submit(); return false;" href="#"/>  ici </a>
-            <input type="hidden" name="csrf_token" value="{generate_csrf()}"/>
-        </form>. <br/>
-        <b>Sans cette information, vous ne pourrez pas participer aux collectives.</b> <br/>
-        Si besoin, vous pouvez contacter le support à 
-        <a href="mailto:{Configuration.SUPPORT_EMAIL}">{Configuration.SUPPORT_EMAIL}</a>.
-        """
+    action = "Récupération " if is_recover else "Création"
+    reason = "récupérer" if is_recover else "créer"
+    form.submit.label.text = f"{reason.capitalize()} le compte"
+    return render_template(
+        "auth/token_confirmation.html",
+        form=form,
+        title=f"{action} de compte",
+    )
